backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
PyYAML==6.0.1
python-magic==0.4.27
Pillow==10.1.0
mutagen==1.47.0
